move `types` out of `apicultor` lib
