--- athena/apicultor/types/__init__.py
@@ -1,4 +1,0 @@
-from athena.apicultor.types.candle import Candle
-from athena.apicultor.types.period import Period
-
-__all__ = ["Candle", "Period"]
