+++ tests/types/test_period.py
@@ -1,6 +1,6 @@
 import pytest
 
-from athena.apicultor.types.period import fill_missing_attributes
+from athena.types.period import fill_missing_attributes
 
 
 @pytest.mark.parametrize(
